Updated file naming structure to fit more modern conventions
