+++ tests/Loaders/JsonLoaderTest.php
@@ -7,99 +7,97 @@
 * @version				1.1.1
 */
 
-require_once 'dice.php';
-require_once 'loader/xml.php';
-require_once 'loader/callback.php';
+require_once 'Dice.php';
+require_once 'Loader/Json.php';
+require_once 'Loader/Callback.php';
 
 
-class XmlLoaderTest extends PHPUnit_Framework_TestCase {
+class JsonLoaderTest extends PHPUnit_Framework_TestCase {
 	private $dice;
-	private $xmlLoader;
+	private $jsonLoader;
 
 	protected function setUp() {
 		parent::setUp ();
 		$this->dice = $this->getMock('\\Dice\\Dice', array('getRule', 'addRule'));		
 		$this->dice->expects($this->any())->method('getRule')->will($this->returnValue(new \Dice\Rule));
-		$this->xmlLoader = new \Dice\Loader\XML;
+		$this->jsonLoader = new \Dice\Loader\Json;
 	}
 
 	protected function tearDown() {
 		$this->dice = null;	
-		$this->xmlLoader = null;
+		$this->jsonLoader = null;
 		parent::tearDown ();
 	}
 	
 	
 	public function testSetDefaultRule() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>*</name>
-		<shared>true</shared>
-	</rule>
-</dice>';
+		$json = '{
+"rules": [				
+		{
+			"name": "*",
+			"shared": true
+		}
+	]
+}';
 		
 	
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->shared = true;
 		
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('*'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 		
 	}
 	
 	public function testShared() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<shared>true</shared>
-	</rule>
-</dice>';
+		$json = '{
+"rules": [				
+		{
+			"name": "A",
+			"shared": true
+		}
+	]
+}';
 		
 		
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->shared = true;
 		
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 	
 	public function testConstructParams() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<construct>
-			<param>A</param>
-			<param>B</param>
-		</construct>
-	</rule>
-</dice>';
-	
+		$json = '{
+"rules": [				
+		{
+			"name": "A",
+			"construct": ["A", "B"]
+		}
+	]
+}';	
 	
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->constructParams[] = 'A';
 		$equivalentRule->constructParams[] = 'B';
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 	public function testSubstitutions() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<substitute>
-			<use>C</use>
-			<as>B</as>
-		</substitute>
-	</rule>
-</dice>';
+		$json = '{
+"rules": [				
+		{
+			"name": "A",
+			"substitute": {"B": "C"}
+		}
+	]
+}';	
+	
 	
 	
 		$equivalentRule = new \Dice\Rule;
@@ -107,27 +105,19 @@ class XmlLoaderTest extends PHPUnit_Framework_TestCase {
 		
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 	public function testSubstitutions2() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<substitute>
-			<use>C</use>
-			<as>B</as>
-		</substitute>
-				
-		<substitute>
-			<use>E</use>
-			<as>F</as>
-		</substitute>
-	</rule>
-</dice>';
-	
+		$json = '{
+"rules": [				
+		{
+			"name": "A",
+			"substitute": {"B": "C", "F": "E"}
+		}
+	]
+}';	
 	
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->substitutions['B'] = new \Dice\Instance('C');
@@ -135,117 +125,116 @@ class XmlLoaderTest extends PHPUnit_Framework_TestCase {
 	
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 	public function testNewInstances() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<newinstance>C</newinstance>
-		<newinstance>D</newinstance>
-		<newinstance>E</newinstance>
-	</rule>
-</dice>';
-		
+		$json = '{
+"rules": [
+		{
+			"name": "A",
+			"newinstances": ["C", "D", "E"]
+		}
+	]
+}';		
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->newInstances = ['C', 'D', 'E'];	
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
+
 	public function testInstanceOf() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>[C]</name>
-		<instanceof>C</instanceof>
-	</rule>
-</dice>';
-	
+		$json = '{
+"rules": [
+		{
+			"name": "[C]",
+			"instanceof": "C"
+		}
+	]
+}';		
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->instanceOf = 'C';
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('[C]'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 	public function testCall() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<call>
-			<method>setFoo</method> 
-           <params> 
-                <param>Foo</param> 
-                <param>Bar</param> 
-            </params>
-		</call> 
-	</rule>
-</dice>';
+		$json = '{
+"rules": [
+		{
+			"name": "A",
+			"call": [
+					["setFoo", ["Foo", "Bar"]]
+				]
+		}
+	]
+}';		
 	
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->call[] = ['setFoo', ['Foo', 'Bar']];
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 
 	public function testInherit() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<inherit>true</inherit>
-	</rule>
-</dice>';
+		$json = '{
+"rules": [
+		{
+			"name": "A",
+			"inherit": true
+		}
+	]
+}';		
+	
 	
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->inherit = true;
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
 	
 	public function testInherit2() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<inherit>false</inherit>
-	</rule>
-</dice>';
-	
+		$json = '{
+"rules": [
+		{
+			"name": "A",
+			"inherit": false
+		}
+	]
+}';		
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->inherit = false;
 	
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
 	
-	public function testShareInstance() {
-		$xml = '<?xml version="1.0"?>
-<dice>
-	<rule>
-		<name>A</name>
-		<shareinstance>C</shareinstance>
-		<shareinstance>D</shareinstance>
-	</rule>
-</dice>';
 	
+	
+	public function testShareInstance() {
+		$json = '{
+"rules": [
+		{
+			"name": "A",
+			"shareinstances": ["C", "D"]
+		}
+	]
+}';		
 		$equivalentRule = new \Dice\Rule;
 		$equivalentRule->shareInstances = ['C', 'D'];
 		
 		$this->dice->expects($this->once())->method('addRule')->with($this->equalTo('A'), $this->equalTo($equivalentRule));
-		$this->xmlLoader->load(simplexml_load_string($xml), $this->dice);
+		$this->jsonLoader->load($json, $this->dice);
 	}
